requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.7.1
google-generativeai==0.8.0
requests==2.31.0
python-dotenv==1.0.0
PyAudio==0.2.11
websockets==12.0
sqlite3-lib==0.1.0
aiohttp==3.9.1
asyncio==3.4.3
json5==0.9.14
pydub==0.25.1
PyYAML==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
